🩹 [Patch]: Encode all PowerShell files using UTF8 with BOM (#2)
## Description

This pull request makes minor changes to the PowerShell files in the
Function App. The only updates are the addition of a Unicode Byte Order
Mark (BOM) at the start of both `run.ps1` and `profile.ps1` files, which
does not affect functionality but may improve compatibility with some
editors and tools.

- Added a BOM to the beginning of `FunctionApp/PostToTeams/run.ps1`
- Added a BOM to the beginning of `FunctionApp/profile.ps1`